% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni16_functions.R
\name{add_additional_groups}
\alias{add_additional_groups}
\title{Add 'all' groups to data frame - annual totals, partnership totals,
Clackmannanshire & Stirling totals, and Soctland totals, then aggregate}
\usage{
add_additional_groups(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame with the additional totals
}
\description{
Add 'all' groups to data frame - annual totals, partnership totals,
Clackmannanshire & Stirling totals, and Soctland totals, then aggregate
}
