% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni15_functions.R
\name{add_additional_groups_ni15}
\alias{add_additional_groups_ni15}
\title{Add additional groups and aggregate to locality level}
\usage{
add_additional_groups_ni15(data)
}
\arguments{
\item{data}{A data frame from \code{\link[=match_smra_and_deaths_ni15]{match_smra_and_deaths_ni15()}}}
}
\value{
A data frame with additional groups
}
\description{
Add additional groups and aggregate to locality level
}
