% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_all_grouping.R
\name{add_all_grouping}
\alias{add_all_grouping}
\title{Add an 'all' grouping to a variable}
\usage{
add_all_grouping(data, group_var, group)
}
\arguments{
\item{data}{The data frame to transform}

\item{group_var}{The name of the variable to transform}

\item{group}{The name of the 'all' grouping}
}
\value{
The data frame with added rows with the 'all' grouping
}
\description{
Adding an 'all' grouping means that, for example, in the case of 'age'
we can use this function to keep the ages in that column but also add rows for 'all ages'.
}
\examples{
tibble::tribble(~age, ~value, "1", 3, "2", 5, "3", 6) \%>\%
  add_all_grouping("age", "All ages")
}
