% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni12_functions.R
\name{add_all_groupings_ni12}
\alias{add_all_groupings_ni12}
\title{Add various 'all' groups for use with NI12}
\usage{
add_all_groupings_ni12(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame with various 'all' groupings
}
\description{
Add various 'all' groups for use with NI12
}
