% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni13_functions.R
\name{add_all_groupings_ni13}
\alias{add_all_groupings_ni13}
\title{Split the data into calendar year and financial year, and add yearly totals,
Scotland totals, locality totals, and C&S totals}
\usage{
add_all_groupings_ni13(data)
}
\arguments{
\item{data}{A data frame after \code{\link[=prepare_for_groupings]{prepare_for_groupings()}}}
}
\value{
A data frame with the described groups added and aggregated
}
\description{
Split the data into calendar year and financial year, and add yearly totals,
Scotland totals, locality totals, and C&S totals
}
