% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_clacks_and_stirling.R
\name{add_clacks_and_stirling}
\alias{add_clacks_and_stirling}
\title{Add rows for Clackmannanshire & Stirling as a partnership}
\usage{
add_clacks_and_stirling(data, council_variable)
}
\arguments{
\item{data}{The data frame containing a column with council area names}

\item{council_variable}{The variable to use for council names}
}
\value{
A data frame with additional rows for C&S
}
\description{
Add rows for Clackmannanshire & Stirling as a partnership
}
