% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni14_functions.R
\name{add_readmission_flag}
\alias{add_readmission_flag}
\title{Add flag for emergency readmissions within 28 days of last admission}
\usage{
add_readmission_flag(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame with a new variable, flag28
}
\description{
An emergency readmission is defined as an admission with
codes 20, 21, 22 or between 30-39. If the person's next admission is within
28 days of the previous one and is an emergency, set \code{flag28} = \code{TRUE}.
}
