% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni14_functions.R
\name{add_smra_death_flag}
\alias{add_smra_death_flag}
\title{Create a flag of deaths from the SMRA discharge type}
\usage{
add_smra_death_flag(discharge_type)
}
\arguments{
\item{discharge_type}{A vector of discharge types. Must be of class \code{integer}}
}
\value{
A boolean vector of death flags
}
\description{
Create a flag of deaths from the SMRA discharge type
}
