% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni16_functions.R
\name{aggregate_to_date_level}
\alias{aggregate_to_date_level}
\title{Aggregate the NI16 SMR extract to a chosen date level}
\usage{
aggregate_to_date_level(data, date_level = c("calendar", "financial"))
}
\arguments{
\item{data}{The SMR extract derived from \code{\link[=process_ni16_smra_extract]{process_ni16_smra_extract()}}}

\item{date_level}{Either \code{calendar} or \code{financial}}
}
\value{
The extract aggregated to chosen date level, with falls as the summary variable
}
\description{
Aggregate the NI16 SMR extract to a chosen date level
}
