% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{calculate_fin_month_from_string}
\alias{calculate_fin_month_from_string}
\title{Given a vector of month names, calculate the financial month in "MX" format}
\usage{
calculate_fin_month_from_string(month_variable)
}
\arguments{
\item{month_variable}{A vector of month names}
}
\value{
A vector of financial months
}
\description{
Given a vector of month names, calculate the financial month in "MX" format
}
