% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{calculate_financial_month}
\alias{calculate_financial_month}
\title{Calculate financial month}
\usage{
calculate_financial_month(date_variable)
}
\arguments{
\item{date_variable}{A vector of dates}
}
\value{
A vector of financial months
}
\description{
Gives financial month in "MX" format, where "M1" is April
of a given year
}
