% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni15_functions.R
\name{calculate_length_of_stay}
\alias{calculate_length_of_stay}
\title{Calculate length of stay (los) for records from SMR01, SMR01e, and SMR04}
\usage{
calculate_length_of_stay(data)
}
\arguments{
\item{data}{The outputs from \code{\link[=process_ni15_smr_extracts]{process_ni15_smr_extracts()}}}
}
\value{
A data frame with length of stay (los)
}
\description{
Calculate length of stay (los) for records from SMR01, SMR01e, and SMR04
}
