% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni13_functions.R
\name{calculate_monthly_beddays}
\alias{calculate_monthly_beddays}
\title{Calculate monthly beddays for NI12}
\usage{
calculate_monthly_beddays(data, year)
}
\arguments{
\item{data}{The SLF after having been aggregated from \code{\link[=prepare_slf_episode_file]{prepare_slf_episode_file()}}}

\item{year}{The year in "XXYY" format}
}
\value{
The data with monthly beddays as variables
}
\description{
Calculate monthly beddays for NI12
}
