% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni12_functions.R
\name{calculate_ni12}
\alias{calculate_ni12}
\title{Calculate NI12 from an SLF episode file aggregated to CIJ level}
\usage{
calculate_ni12(data, fin_year_start, fin_year_end)
}
\arguments{
\item{data}{An episode file aggregated to CIJ level: see \code{\link[=prepare_slf_episode_file]{prepare_slf_episode_file()}}}

\item{fin_year_start}{The start of the financial year}

\item{fin_year_end}{The end of the financial year}
}
\value{
A list of data frames, one at calendar year and one at financial year level
}
\description{
Calculate NI12 from an SLF episode file aggregated to CIJ level
}
