% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni13_functions.R
\name{calculate_ni13}
\alias{calculate_ni13}
\title{Fully calculate NI13 for a given financial year and save out if required}
\usage{
calculate_ni13(year, write_to_disk = FALSE)
}
\arguments{
\item{year}{Financial year in "XXYY" format}

\item{write_to_disk}{When TRUE, writes to disk}
}
\value{
A list of data frames
}
\description{
Fully calculate NI13 for a given financial year and save out if required
}
\seealso{
\code{\link[=get_ni_output_dir]{get_ni_output_dir()}}
}
