% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni14_functions.R
\name{calculate_ni14}
\alias{calculate_ni14}
\title{Use helper functions to calculate the final totals for NI14}
\usage{
calculate_ni14()
}
\value{
A data frame
}
\description{
Use helper functions to calculate the final totals for NI14
}
