% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni15_functions.R
\name{calculate_ni15}
\alias{calculate_ni15}
\title{Create financial year and calendar year totals for NI15}
\usage{
calculate_ni15(extract_start, extract_end)
}
\arguments{
\item{extract_start}{Start of the extract to be taken from SMRA (\%d-\%b-\%Y)}

\item{extract_end}{End of the extract to be taken from SMRA (\%d-\%b-\%Y)}
}
\value{
A list of data frames
}
\description{
Create financial year and calendar year totals for NI15
}
