% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni16_functions.R
\name{calculate_ni16_final_output}
\alias{calculate_ni16_final_output}
\title{Calculate the final NI16 values for calendar year and financial year}
\usage{
calculate_ni16_final_output(write_to_disk = TRUE)
}
\arguments{
\item{write_to_disk}{Boolean parameter, choose to write the output to disk or not}
}
\value{
A list of data frames, one for calendar year named \code{calendar} and one
for financial year named \code{financial}
}
\description{
Calculate the final NI16 values for calendar year and financial year
}
\seealso{
\code{\link[=process_ni16_smra_extract]{process_ni16_smra_extract()}}, \code{\link[=aggregate_to_date_level]{aggregate_to_date_level()}},
\code{\link[=add_additional_groups]{add_additional_groups()}}, \code{\link[=create_population_lookup]{create_population_lookup()}}
}
