% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni18_functions.R
\name{calculate_ni18}
\alias{calculate_ni18}
\title{Read in raw NI18 data and calculate final NI18 output}
\usage{
calculate_ni18(path, min_year, write_to_disk = TRUE)
}
\arguments{
\item{path}{Path to spreadsheet received from Social Care team}

\item{min_year}{Minimum year to use for output}

\item{write_to_disk}{Boolean, whether to write output or not}
}
\value{
A list of two outputs, one for the MI spreadsheet and one for Tableau
}
\description{
Read in raw NI18 data and calculate final NI18 output
}
