% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni19_functions.R
\name{calculate_ni19}
\alias{calculate_ni19}
\title{Calculate quarterly and annual figures for NI19}
\usage{
calculate_ni19(file_from_dd_team, write_to_disk = FALSE)
}
\arguments{
\item{file_from_dd_team}{The name of the file sent from the Delayed Discharges team}

\item{write_to_disk}{When TRUE, will write the output to disk}
}
\value{
The annual figures for NI19
}
\description{
Calculate quarterly and annual figures for NI19
}
\seealso{
\code{\link[=get_ni_input_dir]{get_ni_input_dir()}} and \code{\link[=read_population_lookup]{read_population_lookup()}}
}
