% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_indicator_functions.R
\name{calculate_outcome_indicators}
\alias{calculate_outcome_indicators}
\title{Calculate the data for indicators 1 to 9 using most recent HACE data}
\usage{
calculate_outcome_indicators(
  hace_year,
  sheet_num_hscp = 2,
  sheet_num_loc = 2,
  write_to_disk = FALSE
)
}
\arguments{
\item{hace_year}{The year of the HACE survey}

\item{sheet_num_hscp}{The sheet number in the HSCP Excel workbook to read from}

\item{sheet_num_loc}{The sheet number in the Locality Excel workbook to read from}
}
\value{
A final output for the outcome indicators
}
\description{
Calculate the data for indicators 1 to 9 using most recent HACE data
}
