% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_smra.R
\name{connect_to_smra}
\alias{connect_to_smra}
\title{Code to connect to SMRA database}
\usage{
connect_to_smra()
}
\value{
Formal class Oracle object
}
\description{
You will be asked for a username and password when using this function
}
