% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{convert_fin_month_to_quarter}
\alias{convert_fin_month_to_quarter}
\title{Convert financial month to financial quarter}
\usage{
convert_fin_month_to_quarter(month_variable)
}
\arguments{
\item{month_variable}{A vector of financial months in "MX" format}
}
\value{
A vector of financial quarters in "QX" format
}
\description{
Convert financial month to financial quarter
}
\note{
The return vector will preserve any Annual values
}
\seealso{
Other Output functions: 
\code{\link{join_recalculated_scotland_column}()},
\code{\link{recalculate_quarterly_values}()}
}
\concept{Output functions}
