% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_population_lookup.R
\name{create_lca_population_lookup}
\alias{create_lca_population_lookup}
\title{Create a population lookup for use with NIs at LCA level}
\usage{
create_lca_population_lookup(
  min_year,
  pop_est_path = get_population_estimate_path(),
  spd_path = get_spd_path()
)
}
\arguments{
\item{min_year}{The earliest year to use in the lookup}

\item{pop_est_path}{See \code{\link[=get_population_estimate_path]{get_population_estimate_path()}} or use custom path}

\item{spd_path}{See \code{\link[=get_spd_path]{get_spd_path()}} or use custom path}
}
\value{
A data frame to use as a population lookup for: over 18s, over 65s,
and over 75s
}
\description{
Create a population lookup for use with NIs at LCA level
}
