% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scotland_column.R
\name{create_scotland_column}
\alias{create_scotland_column}
\title{Create an additional column with Scotland values for use in Tableau}
\usage{
create_scotland_column(data)
}
\arguments{
\item{data}{A data frame containing rows where the partnership is Scotland}
}
\value{
A data frame with an additional variable, \code{scotland}, where the values
can be matched on \code{year} and \code{time_period}
}
\description{
Create an additional column with Scotland values for use in Tableau
}
