% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni12_functions.R
\name{flag_admission_records}
\alias{flag_admission_records}
\title{Flag admission records for NI12}
\usage{
flag_admission_records(record_keydate1, fin_year_start, fin_year_end)
}
\arguments{
\item{record_keydate1}{A vector of admission dates}

\item{fin_year_start}{The start of the financial year being calculated}

\item{fin_year_end}{The end of the financial year being calculated}
}
\value{
A boolean vector denoting which records fell within the financial year
}
\description{
Flag admission records for NI12
}
