% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni13_functions.R
\name{format_date_levels_ni13}
\alias{format_date_levels_ni13}
\title{Format the months and year for splitting}
\usage{
format_date_levels_ni13(data, year)
}
\arguments{
\item{data}{A data frame}

\item{year}{The financial year in "XXYY" format}
}
\value{
A data frame with added columns for population year, financial year,
calendar year, calendar month, and financial month
}
\description{
Format the months and year for splitting
}
