% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_indicator_functions.R
\name{format_raw_hace_data}
\alias{format_raw_hace_data}
\title{Format the raw HACE data into the format we need}
\usage{
format_raw_hace_data(data_list, hace_year)
}
\arguments{
\item{data_list}{Output from \code{\link[=read_raw_hace_data]{read_raw_hace_data()}}}

\item{hace_year}{The year of the questionnaire}
}
\value{
A data frame
}
\description{
Format the raw HACE data into the format we need
}
