% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{get_different_years}
\alias{get_different_years}
\title{Get the full financial year, calendar years, and population year from a
single year input}
\usage{
get_different_years(year)
}
\arguments{
\item{year}{A financial year in format "XXYY"}
}
\value{
A named vector in the format c(financial_year, first_cal_year, second_cal_year, pop_year)
}
\description{
Get the full financial year, calendar years, and population year from a
single year input
}
