% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locality_from_postcode.R
\name{get_locality_from_postcode}
\alias{get_locality_from_postcode}
\title{Use postcode to determine the HSCP locality of residence}
\usage{
get_locality_from_postcode(
  data,
  spd_path = get_spd_path(),
  locality_path = get_locality_path()
)
}
\arguments{
\item{data}{Any data frame containing a variable "postcode"}

\item{spd_path}{Path to the SPD lookup, defaults to \code{\link[=get_spd_path]{get_spd_path()}}}

\item{locality_path}{Path to the HSCP locality lookup, defaults to \code{\link[=get_locality_path]{get_locality_path()}}}
}
\value{
A data frame with additional geographies
}
\description{
Use postcode to determine the HSCP locality of residence
}
