% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lookup_paths.R
\name{get_locality_path}
\alias{get_locality_path}
\title{Locality File Path}
\usage{
get_locality_path(ext = "rds")
}
\arguments{
\item{ext}{The extension (type of the file) - optional}

\item{file_name}{(optional) the file name of the Localities files, if not
supplied it will try to return the latest file automatically (using
\code{\link[=find_latest_file]{find_latest_file()}})}
}
\value{
An \code{\link[fs:path]{fs::path()}} to the Scottish Postcode Directory
}
\description{
Get the path to the centrally held HSCP Localities file.
}
