% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni13_functions.R
\name{get_month_intervals}
\alias{get_month_intervals}
\title{Define and return the intervals for each month of the financial year}
\usage{
get_month_intervals(year)
}
\arguments{
\item{year}{A year in format "XXYY"}
}
\value{
A vector of \code{\link[lubridate:interval]{lubridate::interval()}} objects
}
\description{
Define and return the intervals for each month of the financial year
}
