% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni18_functions.R
\name{get_new_ni_18_data}
\alias{get_new_ni_18_data}
\title{Read in and format NI18 data from Social Care team}
\usage{
get_new_ni_18_data(path, min_year)
}
\arguments{
\item{path}{Path to spreadsheet received from Social Care team}

\item{min_year}{Minimum year to use for output}
}
\value{
A data frame
}
\description{
Read in and format NI18 data from Social Care team
}
