% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_paths.R
\name{get_ni_excel_output_dir}
\alias{get_ni_excel_output_dir}
\title{NI Spreadsheet output dir}
\usage{
get_ni_excel_output_dir()
}
\value{
The path to the spreadsheet output folder
}
\description{
NI Spreadsheet output dir
}
