% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_paths.R
\name{get_ni_tableau_output_dir}
\alias{get_ni_tableau_output_dir}
\title{NI Tableau output dir}
\usage{
get_ni_tableau_output_dir()
}
\value{
The path to the Tableau output folder
}
\description{
NI Tableau output dir
}
