% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lookup_paths.R
\name{get_populations_path}
\alias{get_populations_path}
\title{Population Estimates File Path}
\usage{
get_populations_path(file_name = NULL, ext = "rds")
}
\arguments{
\item{file_name}{(optional) the file name of the Population file, if not supplied
it will try to return the latest file automatically (using
\code{\link[=find_latest_file]{find_latest_file()}})}

\item{ext}{The extension (type of the file) - optional}
}
\value{
An \code{\link[fs:path]{fs::path()}} to the Population file
}
\description{
Get the path to the centrally held Population Estimates file
}
