% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lookup_paths.R
\name{get_simd_path}
\alias{get_simd_path}
\title{SIMD File Path}
\usage{
get_simd_path(file_name = NULL, ext = "parquet")
}
\arguments{
\item{file_name}{(optional) the file name of the SIMD file, if not supplied
it will try to return the latest file automatically (using
\code{\link[=find_latest_file]{find_latest_file()}})}

\item{ext}{The extension (type of the file) - optional}
}
\value{
An \code{\link[fs:path]{fs::path()}} to the SIMD file
}
\description{
Get the path to the centrally held Scottish Index of Multiple
Deprivation (SIMD) file.
}
