% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lookup_paths.R
\name{get_spd_path}
\alias{get_spd_path}
\title{Scottish Postcode Directory File Path}
\usage{
get_spd_path(file_name = NULL, ext = "parquet")
}
\arguments{
\item{file_name}{(optional) the file name of the SPD, if not supplied it will
try to return the latest file automatically (using \code{\link[=find_latest_file]{find_latest_file()}})}

\item{ext}{The extension (type of the file) - optional}
}
\value{
An \code{\link[fs:path]{fs::path()}} to the Scottish Postcode Directory
}
\description{
Get the path to the centrally held Scottish Postcode Directory
(SPD) file.
}
