% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-update-refs.R
\name{indicators_to_update}
\alias{indicators_to_update}
\title{Vector of the indicators being updated this run}
\usage{
indicators_to_update()
}
\value{
A character vector of indicators, in format "NIXX"
}
\description{
Vector of the indicators being updated this run
}
\seealso{
Other Initialisation: 
\code{\link{latest_hace_year}()},
\code{\link{latest_update}()},
\code{\link{previous_update}()},
\code{\link{slf_indicator_update_periods}()}
}
\concept{Initialisation}
