% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{join_recalculated_scotland_column}
\alias{join_recalculated_scotland_column}
\title{Join a recalculated Scotland column to the indicator data frame}
\usage{
join_recalculated_scotland_column(data)
}
\arguments{
\item{data}{A data frame with recalculated indicator values}
}
\value{
A data frame with the Scotland column replaced with accurate values
}
\description{
Join a recalculated Scotland column to the indicator data frame
}
\seealso{
Other Output functions: 
\code{\link{convert_fin_month_to_quarter}()},
\code{\link{recalculate_quarterly_values}()}
}
\concept{Output functions}
