% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-update-refs.R
\name{latest_hace_year}
\alias{latest_hace_year}
\title{Latest year of the HACE indicators}
\usage{
latest_hace_year()
}
\value{
latest HACE year as YYYY
}
\description{
Get the date of the latest HACE year
}
\seealso{
Other Initialisation: 
\code{\link{latest_update}()},
\code{\link{previous_update}()}
}
\concept{Initialisation}
