% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-update-refs.R
\name{latest_update}
\alias{latest_update}
\title{Latest update}
\usage{
latest_update()
}
\value{
Latest update as MMM-YYYY
}
\description{
Get the date of the latest update, e.g 'Sep-2022'
}
\seealso{
Other Initialisation: 
\code{\link{indicators_to_update}()},
\code{\link{latest_hace_year}()},
\code{\link{previous_update}()},
\code{\link{slf_indicator_update_periods}()}
}
\concept{Initialisation}
