% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni14_functions.R
\name{match_smra_and_deaths}
\alias{match_smra_and_deaths}
\title{Match SMRA extract to NRS deaths extract to ensure data quality and
removal of records where individual died}
\usage{
match_smra_and_deaths(smra_data, nrs_data)
}
\arguments{
\item{smra_data}{Extract created by \link{process_ni14_smra_extract}}

\item{nrs_data}{Extract created by \link{process_ni14_nrs_extract}}
}
\value{
A matched data frame with additional variables, key variable being 'stay',
which is TRUE when individual is not dead
}
\description{
Match SMRA extract to NRS deaths extract to ensure data quality and
removal of records where individual died
}
