% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni15_functions.R
\name{match_smra_and_deaths_ni15}
\alias{match_smra_and_deaths_ni15}
\title{Match the deaths data and the SMR data, and split into calendar and financial year}
\usage{
match_smra_and_deaths_ni15(smr_data, deaths_data)
}
\arguments{
\item{smr_data}{SMR data with length of stay calculated}

\item{deaths_data}{GRO data}
}
\value{
A list of data frames
}
\description{
Match the deaths data and the SMR data, and split into calendar and financial year
}
