% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{ni14_gro_query}
\alias{ni14_gro_query}
\title{Query to extract NI14 deaths data from SMRA}
\usage{
ni14_gro_query()
}
\value{
An SQL query as a text string
}
\description{
Query to extract NI14 deaths data from SMRA
}
