% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{ni16_smra_query}
\alias{ni16_smra_query}
\title{Query to extract NI16 data from SMRA}
\usage{
ni16_smra_query()
}
\value{
An SQL query as a text string
}
\description{
Query to extract NI16 data from SMRA
}
