% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsni-package.R
\docType{package}
\name{phsni-package}
\alias{phsni}
\alias{phsni-package}
\title{phsni: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Public-Health-Scotland/national-indicators}
  \item Report bugs at \url{https://github.com/Public-Health-Scotland/national-indicators/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
