% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni13_functions.R
\name{prepare_for_groupings}
\alias{prepare_for_groupings}
\title{Match on geographies to data, select only relevant variables and aggregate for NI13}
\usage{
prepare_for_groupings(data)
}
\arguments{
\item{data}{A data frame after \code{\link[=format_date_levels_ni13]{format_date_levels_ni13()}}}
}
\value{
An aggregated data frame to Locality level at the lowest
}
\description{
Match on geographies to data, select only relevant variables and aggregate for NI13
}
