% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hace_outputs.R
\name{process_hace_outputs}
\alias{process_hace_outputs}
\title{Process the HACE outputs and save the outputs}
\usage{
process_hace_outputs(finyear, archive = TRUE, write_to_disk = TRUE)
}
\arguments{
\item{finyear}{The financial year to be shown in the column Year1, in "20XX/YY" format}

\item{archive}{Default to \code{TRUE}, when TRUE then will archive the previous NI1-9 Tableau output}

\item{write_to_disk}{Defaults to \code{TRUE}. Writes three files to disk, the 'all data and variables'
version, the Tableau output, and the Management Information output}
}
\value{
The final, unfiltered, NI1-9 output
}
\description{
Process the HACE outputs and save the outputs
}
