% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni14_functions.R, R/process_ni14_gro_extract.R
\name{process_ni14_gro_extract}
\alias{process_ni14_gro_extract}
\title{Read in and process GRO Deaths extract for NI14}
\usage{
process_ni14_gro_extract()

process_ni14_gro_extract()
}
\value{
A data frame

A data frame
}
\description{
Read in and process GRO Deaths extract for NI14

Read in and process GRO Deaths extract for NI14
}
