% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni14_functions.R
\name{process_ni14_nrs_extract}
\alias{process_ni14_nrs_extract}
\title{Read in and process GRO Deaths extract for NI14}
\usage{
process_ni14_nrs_extract(min_date = "01-APR-2022")
}
\arguments{
\item{min_date}{The minimum date for date of death, read from gro}
}
\value{
A data frame of death dates and link_no
}
\description{
Read in and process GRO Deaths extract for NI14
}
