% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ni15_gro_extract.R
\name{process_ni15_gro_extract}
\alias{process_ni15_gro_extract}
\title{Process NRS deaths extract for use with NI15}
\usage{
process_ni15_gro_extract(extract_start, extract_end)
}
\arguments{
\item{extract_start}{Start date of extract in 'DD-MMM-YYYY' format}

\item{extract_end}{End date of extract in 'DD-MMM-YYYY' format}
}
\value{
A tibble
}
\description{
Process NRS deaths extract for use with NI15
}
\seealso{
"SQL/ni15_gro.sql"
}
