% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ni15_smr_extracts.R
\name{process_ni15_smr_extracts}
\alias{process_ni15_smr_extracts}
\title{Process SMR01 and SMR01e data for use with NI15}
\usage{
process_ni15_smr_extracts(extract_start, extract_end)
}
\arguments{
\item{extract_start}{Start date of extract in 'DD-MMM-YYYY' format}

\item{extract_end}{End date of extract in 'DD-MMM-YYYY' format}
}
\value{
A tibble
}
\description{
Process SMR01 and SMR01e data for use with NI15
}
