% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni16_functions.R
\name{process_ni16_smra_extract}
\alias{process_ni16_smra_extract}
\title{Read the SMRA extract for NI16, filter to records involving falls,
and match on geographies}
\usage{
process_ni16_smra_extract(min_date = "01-APR-2013")
}
\value{
A data frame
}
\description{
Read the SMRA extract for NI16, filter to records involving falls,
and match on geographies
}
\seealso{
\code{\link[=connect_to_smra]{connect_to_smra()}}, \code{\link[=calculate_financial_month]{calculate_financial_month()}}, \code{\link[=get_spd_path]{get_spd_path()}},
\code{\link[=get_locality_path]{get_locality_path()}}
}
