% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_population_lookup.R
\name{read_population_lookup}
\alias{read_population_lookup}
\title{Read the population lookup if one exists, otherwise create one}
\usage{
read_population_lookup(
  min_year,
  update_suffix = latest_update(),
  ages_required = c("over18", "over65", "over75")
)
}
\arguments{
\item{min_year}{The minimum year to be passed to \code{\link[=create_population_lookup]{create_population_lookup()}}}

\item{update_suffix}{Defaults to \code{\link[=latest_update]{latest_update()}}, the suffix of the population lookup}

\item{ages_required}{One of "over18", "over65", "over75", selects ony relevant columns}
}
\value{
The population lookup, either read from disk or created
}
\description{
Read the population lookup if one exists, otherwise create one
}
