% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_indicator_functions.R
\name{read_raw_hace_data}
\alias{read_raw_hace_data}
\title{Read in raw HACE data for manipulation}
\usage{
read_raw_hace_data(hace_year, sheet_num_hscp, sheet_num_loc)
}
\arguments{
\item{hace_year}{The year of the published HACE results}

\item{sheet_num_hscp}{The Excel sheet number relevant to the HSCP-level data}

\item{sheet_num_loc}{The Excel sheet number relevant to the Locality-level data}
}
\value{
A list of length \code{2}, with the HSCP then the Locality data
}
\description{
Read in raw HACE data for manipulation
}
