% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_slf_episode_ni.R
\name{read_slf_episode_ni}
\alias{read_slf_episode_ni}
\title{Read the SLF episode file for use with the NIs}
\usage{
read_slf_episode_ni(
  fin_year,
  location = c("production", "development"),
  file_type = c("fst", "parquet")
)
}
\arguments{
\item{fin_year}{Financial year in "XXYY" format}

\item{location}{"production" will return /hscdiip and "development" will return /sourcedev}

\item{file_type}{Either "fst" or "parquet" based on which is being read in}
}
\value{
The SLF episode file
}
\description{
Read the SLF episode file for use with the NIs
}
