% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{recalculate_quarterly_values}
\alias{recalculate_quarterly_values}
\title{Recalculate monthly indicator values into quarterly values}
\usage{
recalculate_quarterly_values(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame with Value and Denominator recalculated on a quarterly basis
}
\description{
Recalculate monthly indicator values into quarterly values
}
\seealso{
\code{\link[=convert_fin_month_to_quarter]{convert_fin_month_to_quarter()}}

Other Output functions: 
\code{\link{convert_fin_month_to_quarter}()},
\code{\link{join_recalculated_scotland_column}()}
}
\concept{Output functions}
