% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-update-refs.R
\name{slf_indicator_update_periods}
\alias{slf_indicator_update_periods}
\title{List of the update years and quarters for indicators 12, 13 and 20}
\usage{
slf_indicator_update_periods()
}
\value{
A list of two vectors: \describe{
\item{years}{A vector of financial years in "20XX/YY" format}
\item{quarters}{A vector of financial quarters in "20XX/YY QZ" format}}
}
\description{
List of the update years and quarters for indicators 12, 13 and 20
}
\seealso{
Other Initialisation: 
\code{\link{indicators_to_update}()},
\code{\link{latest_hace_year}()},
\code{\link{latest_update}()},
\code{\link{previous_update}()}
}
\concept{Initialisation}
