% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_partnership_names.R
\name{standardise_partnership_names}
\alias{standardise_partnership_names}
\title{Standardise partnership names for NIs}
\usage{
standardise_partnership_names(partnership_name)
}
\arguments{
\item{partnership_name}{A vector of existing poartnership names}
}
\value{
A vector of partnership names in the NI expected format
}
\description{
Standardise partnership names for NIs
}
