% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni12_functions.R
\name{calculate_ni12}
\alias{calculate_ni12}
\title{Calculate NI12 from an SLF episode file aggregated to CIJ level}
\usage{
calculate_ni12(year, write_to_disk = FALSE)
}
\arguments{
\item{year}{The year of the SLF to extract in "XXYY" format}

\item{wirte_to_disk}{When TRUE, will write the outputs to directory specified in
\code{\link[=get_ni_output_dir]{get_ni_output_dir()}}}
}
\value{
A list of data frames, one at calendar year and one at financial year level
}
\description{
Calculate NI12 from an SLF episode file aggregated to CIJ level
}
