% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_functions.R
\name{format_financial_year}
\alias{format_financial_year}
\title{Convert four-character year into "20XX/YY"}
\usage{
format_financial_year(four_char_year)
}
\arguments{
\item{four_char_year}{A vector of years, in "CCYY" format}
}
\value{
A vector of financial years in "20XX/YY" format
}
\description{
Convert four-character year into "20XX/YY"
}
