% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ni14_smra_extract.R
\name{process_ni14_smra_extract}
\alias{process_ni14_smra_extract}
\title{Read in and process SMRA extract for NI14}
\usage{
process_ni14_smra_extract()
}
\value{
A tibble
}
\description{
Read in and process SMRA extract for NI14
}
