% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_population_lookup.R
\name{read_population_lookup}
\alias{read_population_lookup}
\title{Read the population lookup if one exists, otherwise create one}
\usage{
read_population_lookup(
  ages_required = c("over18", "over65", "over75"),
  type = c("locality", "partnership")
)
}
\arguments{
\item{ages_required}{One of "over18", "over65", "over75", selects ony relevant columns}

\item{type}{The level you need the populations at, either "locality" or "partnership"}
}
\value{
The population lookup, either read from disk or created
}
\description{
Read the population lookup if one exists, otherwise create one
}
